Store DPI scale as a float instead of a double

+++ include/sdlmain.h
@@ -203,8 +203,9 @@ struct SDL_Block {
 
 		PixelFormat pixel_format = {};
 
-		double dpi_scale = 1.0;
-		bool fullscreen  = false;
+		float dpi_scale = 1.0f;
+
+		bool fullscreen = false;
 
 		// This flag indicates, that we are in the process of switching
 		// between fullscreen or window (as oppososed to changing
